jwt verif

+++ app/Controllers/Http/AddressesController.ts
@@ -17,13 +17,19 @@ export default class AddressesController {
      */
     public async index ({response,request}:HttpContextContract){
         try{
-            const user = await User.findOrFail(jwt.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id'])
-            if (user.role.role_id == 5 || user.role.role_id == 4){
-                const addresses = await Address.query()
-                return response.status(200).json({addresses})
+            const user_id = this.getId(request)
+            if (user_id){
+                const user = await User.findOrFail(user_id)
+                if (user.role.role_id == 5 || user.role.role_id == 4){
+                    const addresses = await Address.query()
+                    return response.status(200).json({addresses})
+                }else{
+                    return response.status(403).json({message: "Error you don't have the correct permissions"})
+                }
             }else{
-                return response.status(403).json({message: "Error you don't have the correct permissions"})
+                return response.status(500).json({err:"jwt token error"})
             }
+
 
         }catch(err){
             return response.status(500).json({err})
@@ -41,31 +47,39 @@ export default class AddressesController {
      */
     public async getById ({params,response,request}:HttpContextContract){
         try{
-            const user = await User.findOrFail(jwt.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id'])
-            if (params.type=='payment' && user.fk_payment_address_id!=null){
-                const address = await Address.findOrFail(user.fk_payment_address_id);
-                return response.status(200).json({address})
-            }else if(user.fk_payment_address_id==null){
-                return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
+            const user_id = this.getId(request)
+            if (user_id){
+                const user = await User.findOrFail(user_id)
+                if (params.type=='payment' && user.fk_payment_address_id!=null){
+                    const address = await Address.findOrFail(user.fk_payment_address_id);
+                    return response.status(200).json({address})
+                }else if(user.fk_payment_address_id==null){
+                    return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
+                }
+
+                if(params.type=='delivery' && user.fk_delivery_address_id!=null){
+                    const address = await Address.findOrFail(user.fk_payment_address_id);
+                    return response.status(200).json({address})
+                }else if(user.fk_delivery_address_id==null){
+                    return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
+                }
+
+                if(params.type=='restorer' && user.fk_restorer_id!=null){
+                    const address = await Address.findOrFail(user.restorer.fk_address_id);
+                    return response.status(200).json({address})
+                }else if(user.fk_restorer_id==null){
+                    return response.status(400).json({message : 'This user does not have a restorer yet create it instead'})
+                }
+                if(!user){
+                    return response.status(403).json({message:'error wrong user id'})
+                }
+                else{
+                    return response.status(403).json({message:'error wrong user id'})
+                }
+            }else{
+                return response.status(500).json({err:"jwt token error"})
             }
 
-            if(params.type=='delivery' && user.fk_delivery_address_id!=null){
-                const address = await Address.findOrFail(user.fk_payment_address_id);
-                return response.status(200).json({address})
-            }else if(user.fk_delivery_address_id==null){
-                return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
-            }
-
-            if(params.type=='restorer' && user.fk_restorer_id!=null){
-                const address = await Address.findOrFail(user.restorer.fk_address_id);
-                return response.status(200).json({address})
-            }else if(user.fk_restorer_id==null){
-                return response.status(400).json({message : 'This user does not have a restorer yet create it instead'})
-            }
-
-            else{
-                return response.status(403).json({message:'error wrong user id'})
-            }
         }catch(err){
             return response.status(500).json({err})
         }
@@ -76,26 +90,32 @@ export default class AddressesController {
     public async create({ request , response,params }:HttpContextContract){
 
         try {
-            const user = await User.findOrFail(jwt.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id'])
+            const user_id = this.getId(request)
+            if (user_id){
+                const user = await User.findOrFail(user_id)
+                if (params.type=='payment' && user.fk_payment_address_id==null){
+                    const address = await Address.create({address_city: request.body()['address_city'], address_street: request.body()['address_street'], address_street_number: request.body()['address_street_number'], address_postal_code:request.body()['address_postal_code'] });
+                    await user.related('payment_address_id').associate(address)
+                    return response.status(201).json({address})
+                }else if(user.fk_payment_address_id!=null){
+                    return response.status(400).json({message : 'This user has already a payment address update it instead'})
+                }
 
-            if (params.type=='payment' && user.fk_payment_address_id==null){
-                const address = await Address.create({address_city: request.body()['address_city'], address_street: request.body()['address_street'], address_street_number: request.body()['address_street_number'], address_postal_code:request.body()['address_postal_code'] });
-                await user.related('payment_address_id').associate(address)
-                return response.status(201).json({address})
-            }else if(user.fk_payment_address_id!=null){
-                return response.status(400).json({message : 'This user has already a payment address update it instead'})
+                if(params.type=='delivery' && user.fk_delivery_address_id==null){
+                    const address = await Address.create({address_city: request.body()['address_city'], address_street: request.body()['address_street'], address_street_number: request.body()['address_street_number'], address_postal_code:request.body()['address_postal_code'] });
+                    await user.related('delivery_address_id').associate(address)
+                    return response.status(201).json({address})
+                }else if(user.fk_delivery_address_id!=null){
+                    return response.status(400).json({message : 'This user has already a delivery address update it instead'})
+                }
+                else{
+                    return response.status(500).json({message:'error wrong user id or params'})
+                }
+            }else{
+                return response.status(500).json({err:"jwt token error"})
             }
 
-            if(params.type=='delivery' && user.fk_delivery_address_id==null){
-                const address = await Address.create({address_city: request.body()['address_city'], address_street: request.body()['address_street'], address_street_number: request.body()['address_street_number'], address_postal_code:request.body()['address_postal_code'] });
-                await user.related('delivery_address_id').associate(address)
-                return response.status(201).json({address})
-            }else if(user.fk_delivery_address_id!=null){
-                return response.status(400).json({message : 'This user has already a delivery address update it instead'})
-            }
-            else{
-                return response.status(500).json({message:'error wrong user id or params'})
-            }
+
 
         } catch(err){
             return response.status(500).json({err})
@@ -106,39 +126,42 @@ export default class AddressesController {
 
     public async update({request, response, params}:HttpContextContract){
         try {
-            const user = await User.findOrFail(jwt.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id'])
+            const user_id = this.getId(request)
+            if (user_id){
+                const user = await User.findOrFail(user_id)
+                if (params.type=='payment' && user.fk_payment_address_id!=null){
+                    const address = await Address.findOrFail(user.fk_payment_address_id);
+                    address.merge(request.body());
+                    await address.save();
+                    return response.status(200).json({address})
+                }else if(user.fk_payment_address_id==null){
+                    return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
+                }
 
-            if (params.type=='payment' && user.fk_payment_address_id!=null){
-                const address = await Address.findOrFail(user.fk_payment_address_id);
-                address.merge(request.body());
-                await address.save();
-                return response.status(200).json({address})
-            }else if(user.fk_payment_address_id==null){
-                return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
+                if(params.type=='delivery' && user.fk_delivery_address_id!=null){
+                    const address = await Address.findOrFail(user.fk_payment_address_id);
+                    address.merge(request.body());
+                    await address.save();
+                    return response.status(200).json({address})
+                }else if(user.fk_delivery_address_id==null){
+                    return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
+                }
+
+                if(params.type=='restorer' && user.fk_restorer_id!=null){
+                    const address = await Address.findOrFail(user.restorer.fk_address_id);
+                    address.merge(request.body());
+                    await address.save();
+                    return response.status(200).json({address})
+                }else if(user.fk_restorer_id==null){
+                    return response.status(400).json({message : 'This user does not have a restorer yet create it instead'})
+                }
+
+                else{
+                    return response.status(403).json({message:'Error wrong user id'})
+                }
+            }else{
+                return response.status(500).json({err:"jwt token error"})
             }
-
-            if(params.type=='delivery' && user.fk_delivery_address_id!=null){
-                const address = await Address.findOrFail(user.fk_payment_address_id);
-                address.merge(request.body());
-                await address.save();
-                return response.status(200).json({address})
-            }else if(user.fk_delivery_address_id==null){
-                return response.status(400).json({message : 'This user does not have a payment address yet create it instead'})
-            }
-
-            if(params.type=='restorer' && user.fk_restorer_id!=null){
-                const address = await Address.findOrFail(user.restorer.fk_address_id);
-                address.merge(request.body());
-                await address.save();
-                return response.status(200).json({address})
-            }else if(user.fk_restorer_id==null){
-                return response.status(400).json({message : 'This user does not have a restorer yet create it instead'})
-            }
-
-            else{
-                return response.status(403).json({message:'Error wrong user id'})
-            }
-
         }catch(err){
             return response.status(500).json({err})
         }
@@ -149,41 +172,55 @@ export default class AddressesController {
 
     public async delete({response,request, params}:HttpContextContract){
         try{
+            const user_id = this.getId(request)
+            if (user_id){
+                const user = await User.findOrFail(user_id)
+                if (params.type=='payment' && user.fk_payment_address_id!=null){
+                    const address = await Address.findOrFail(user.fk_payment_address_id);
+                    await address.delete();
+                    return response.status(200).json({address})
+                }else if(user.fk_payment_address_id==null){
+                    return response.status(400).json({message : 'This user does not have a payment address yet'})
+                }
 
-            const user = await User.findOrFail(jwt.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id'])
+                if(params.type=='delivery' && user.fk_delivery_address_id!=null){
+                    const address = await Address.findOrFail(user.fk_payment_address_id);
+                    await address.delete();
+                    return response.status(200).json({address})
+                }else if(user.fk_delivery_address_id==null){
+                    return response.status(400).json({message : 'This user does not have a payment address yet'})
+                }
 
-            if (params.type=='payment' && user.fk_payment_address_id!=null){
-                const address = await Address.findOrFail(user.fk_payment_address_id);
-                await address.delete();
-                return response.status(200).json({address})
-            }else if(user.fk_payment_address_id==null){
-                return response.status(400).json({message : 'This user does not have a payment address yet'})
-            }
+                if(params.type=='restorer' && user.fk_restorer_id!=null){
+                    const address = await Address.findOrFail(user.restorer.fk_address_id);
+                    await address.delete();
+                    return response.status(200).json({address})
+                }else if(user.fk_restorer_id==null){
+                    return response.status(400).json({message : 'This user does not have a restorer yet'})
+                }
 
-            if(params.type=='delivery' && user.fk_delivery_address_id!=null){
-                const address = await Address.findOrFail(user.fk_payment_address_id);
-                await address.delete();
-                return response.status(200).json({address})
-            }else if(user.fk_delivery_address_id==null){
-                return response.status(400).json({message : 'This user does not have a payment address yet'})
-            }
-
-            if(params.type=='restorer' && user.fk_restorer_id!=null){
-                const address = await Address.findOrFail(user.restorer.fk_address_id);
-                await address.delete();
-                return response.status(200).json({address})
-            }else if(user.fk_restorer_id==null){
-                return response.status(400).json({message : 'This user does not have a restorer yet'})
-            }
-
-            else{
-                return response.status(403).json({message:'Error wrong user id'})
+                else{
+                    return response.status(403).json({message:'Error wrong user id'})
+                }
+            }else{
+                return response.status(500).json({err:"jwt token error"})
             }
 
         }catch(err){
             return response.status(500).json({err})
         }
 
+
+    }
+
+    public  getId(request){
+        const token =  request.header('authorization')?.split(" ")
+        try{
+            const user_id = jwt.verify(token[1], 'TOKEN_PRIVATE_KEY')['user_id']
+            return user_id
+        }catch{
+            return false
+        }
 
     }
 }
